requirements.txt
colorama==0.3.9
matplotlib==1.5.1
Pillow==3.1.2
